+++ app/src/main/kotlin/my/github/MrxSiN/pixellauncherevolved/PixelLauncherEvolvedModule.kt
@@ -27,7 +27,7 @@ class PixelLauncherEvolvedModule : XposedModule() {
         val logger = AndroidLogger
         logger.info("Loading in ${param.packageName}")
 
-        FeatureRegistry.installEnabled(
+        FeatureRegistry.install(
             FeatureContext(
                 xposed = this,
                 classLoader = param.defaultClassLoader,